Added update trajectory
update_trajectory now changes the angle of the ball based on where on the
paddle it hit.

Need to get W and H to take into account the bar heights.  Time to add more constants.

Idea: Add  a "sense" private var to the paddles that is set to 1 (for left) or -1 (for right).  This can then be used
to distinguish between left and right paddle to simplify code.