Better AI
I implemented better AI so that it can handle multiple bounces off the walls.  The design
(see fig. 4) just use % and / to find the correct position.  It worked almost perfectly
the first time but I didn't take into account the fact the a) C++ mod function returns a negative number if one of the
operands are negative (it does't on my calculator), and b) I wasn't taking into account the width of the
top bar, which caused an accumulation of error.

I debugged these issues by creating a "target" rectangle of where the calculations predicted the ball
would end up and then just visually comparing them with what really happens.  Another small error remanes, in that
the calculations are based on the "mid" of the paddle and ball, but he acutally intersect calculations are
based on the edges touching.  No big deal.

I also noticed that the lower the incident angle the ball's bounce is to the wall, the bigger the error in the
calculations.  Again this probably has to do with using mid vs edge for the ball, plus error with intersection.

The target has been left in for now and will be removed.