Introducing pong_object
Since all the objects are so similar I want to create a pong_object that
inherits sf::RectangleShape, and then adds the union of data members needed for
both the ball and the paddles.  Then all the functions can be written in terms
of this type.  May simplify a lot.  Let's see.

It seems to work fine for paddle_type for now.