paddle_type is now a pong_rect
Renamed pong_object to pong_rect.

Added functions for mid and to_rect in terms of pong_rect and replaced
paddle_type.

Notice (line 18) how we can convert a member type to an inherited type by temporarily
implementing both.  This would be a pretty bad bug if it wasn't temporary, since the
copy constructor would copy the the wrong *this when initializing shape!